requirements.txt
google-adk==1.19.0
google-genai
pytest==9.0.1
pytest-asyncio==1.3.0
deprecated
pandas
tabulate
tqdm
scikit-learn
google-generativeai  # For Gemini integration
# Additional dependencies for multi-agent and custom tools
pydantic

